bugfix: slaves would send version as a 32-bit integer rather than 64-bit, meaning the major-version was always interpreted as 0.

+++ include/redux/network/host.hpp
@@ -34,7 +34,7 @@ namespace redux {
             
             struct HostInfo {       // "static" information, only exchanged on connect.
                 char littleEndian;
-                int reduxVersion;
+                uint64_t reduxVersion;
                 pid_t pid;
                 uint8_t peerType;
                 uint16_t nCores;
